fix(flag-anime): using faulty cache compare logic
Because the anime ID and anime library IDs are different, we
can't use them to check for existence in our cache; however
using the English title is worse cause it may not exist at all.

The slug must be unique because it refers to the anime page
on the site.